+++ changelog.md
@@ -22,11 +22,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lazy load language switcher
 - Lazy load loader in default layout
 - Added `VueObserveVisibility` dependency
-### Changed
+
+### Changed / Improved
+
 - Updated Vuex to 3.5.1
 - Refactor of some parts of views
 - Loading `homepage` vuex module only when needed (on 2 views)
 - Functional LoaderScoped.vue
+- Remove import phoneNum validator from shipping and payment
 
 ## [1.12.2] - UNRELEASED